Added a new user-config option to change the type and value of the active-page links and their active-parent links:
```
activeLink: {
  type: 'class',
  activeState: 'active',
  parentState: 'active-parent',
}
```

The values listed are the defaults and may be omitted from the config file if you choose not to modify their values.

The allowed values for `type` are: `attr`, `attribute`, and `class`. Any other value will default to setting a `[class]` attribute.

State values are automatically converted to `Kebab` case: `My Active State` becomes `my-active-state`.

For attributes, the attribute is automatically prefixed with `data-` for compliance. For example, if you set the type to `attr` and the `activeState` value to `active-state`,
the resulting link will have an attribute of `<a href="..." data-active-state>`.